move ChartsThemeProvider from ViewDashboard to app level
